Revert "refactor: introduce API versioning structure under /api/v1"
