+++ backend/src/index.ts
@@ -2,9 +2,13 @@
 import express, { Request, Response } from "express";
 import cors from "cors";
 import "dotenv/config";
-import v1Routes from "./routes/v1";
 
-
+import ideasRoutes from "./routes/ideas.routes";
+import experimentsRoutes from "./routes/experiments.routes";
+import outcomesRoutes from "./routes/outcomes.routes";
+import reflectionsRoutes from "./routes/reflections.routes";
+import authRoutes from "./routes/auth.routes";
+import insightsRoutes from "./routes/insights.routes";
 
 // import prisma from "./lib/prisma";
 console.log("INDEX TS SERVER STARTED");
@@ -15,12 +19,15 @@ app.use(cors({ origin: "http://localhost:3000" }));
 app.use(express.json());
 
 app.get("/health", (_req: Request, res: Response) => {
-  
   res.json({ success: true, message: "Backend is running" });
 });
-app.use("/api/v1", v1Routes);
 
-
+app.use("/auth", authRoutes);
+app.use("/ideas", ideasRoutes);
+app.use("/experiments", experimentsRoutes);
+app.use("/outcomes", outcomesRoutes);
+app.use("/reflections", reflectionsRoutes);
+app.use("/insights", insightsRoutes);
 
 app.use((err: any, _req: Request, res: Response, _next: any) => {
   console.error("Unhandled error:", err);
