+++ backend/src/routes/auth.routes.ts
@@ -1,4 +1,3 @@
-
 // backend/src/routes/auth.routes.ts
 import { Router, Request, Response } from "express";
 import { registerUser, loginUser, refreshAccessToken, logoutUser } from "../services/auth.service";
