+++ backend/src/routes/ideas.routes.ts
@@ -1,4 +1,3 @@
-
 import { Router } from "express";
 import {
   getIdeas,
