+++ backend/src/routes/outcomes.routes.ts
@@ -1,4 +1,3 @@
-
 import { Router, Request, Response } from "express";
 
 import {
